V8 - Municipios normalizados

+++ invoiceSave.asp
@@ -1,30 +1,53 @@
 <!--#include file="connect.asp"-->
 <!--#include file="Classes/ValidaProcedimentoProfissional.asp"-->
 <!--#include file="Classes/Logs.asp"-->
+<!--#include file="modulos/audit/AuditoriaUtils.asp"-->
 <%
 
 
 InvoiceID = req("I")
 CD = ref("T")
 msgExtra=""
+statusAlertSuccess="success"
+
+Voucher = ref("voucher")
 temregradesconto=ref("temregradesconto")
 if temregradesconto="" then
+    temregradesconto="0"
+end if
+if Voucher<>"" then
     temregradesconto="0"
 end if
 
 idUsuariosDesconto = "0"
 
+PossuiPermissaoParaDesconto = True
 totalValorDescontado = 0
 totalValorDescontadoEnvio = 0
 totalValorAcrescido = 0
-contaRatiada = 0
+contaRateada = 0
 sqlRateio = "select count(*) total from invoice_rateio where InvoiceID = " & InvoiceID
 set rsRateio = db.execute(sqlRateio)
 if not rsRateio.eof then
 	if ccur(rsRateio("total")) > 0 then
-		contaRatiada = 1
+		contaRateada = 1
 	end if
 end if
+
+
+' ######################### BLOQUEIO FINANCEIRO ########################################
+if verificaBloqueioConta(1, 1, 1, ref("CompanyUnitID"),ref("sysDate")) then
+ %>
+         showMessageDialog("Esta conta está BLOQUEADA e não pode ser alterada!");
+         $("#btnSave").prop("disabled", false);
+
+     saveExecutados();
+ <%
+        response.write(retorno)
+        response.end
+end if
+
+' #####################################################################################
 
 if temregradesconto=1 then
 	'Validar se existe algum desconto cadastrado para o sistema
@@ -47,8 +70,8 @@ if temregradesconto=1 then
 		set rsDescontosUsuario = db.execute("select suser.id as idUser, rd.id, Recursos, Unidades, rd.RegraID, Procedimentos, DescontoMaximo, TipoDesconto "&_
 											" from regrasdescontos rd  "&_
 											" INNER JOIN regraspermissoes rp ON rp.id = rd.RegraID "&_
-											" INNER JOIN sys_users suser on suser.Permissoes LIKE CONCAT('%[',rd.RegraID,']%') "&_
-											" WHERE  rd.Recursos LIKE '%"&querydesconto&"%' AND (rd.Unidades LIKE '%|"& session("UnidadeID") &"|%' OR rd.Unidades  = '' OR rd.Unidades IS NULL OR rd.Unidades  = '0' ) AND rd.RegraID IS NOT NULL")
+											" INNER JOIN sys_users suser on suser.RegraID = rd.RegraID "&_
+											" WHERE rd.Recursos LIKE '%"&querydesconto&"%' AND (rd.Unidades LIKE '%|"& session("UnidadeID") &"|%' OR rd.Unidades  = '' OR rd.Unidades IS NULL OR rd.Unidades  = '0' ) AND rd.RegraID IS NOT NULL")
 
 		'select suser.id, rd.id, Recursos, Unidades, rd.RegraID, Procedimentos, DescontoMaximo, TipoDesconto from regrasdescontos rd inner join sys_users suser on suser.Permissoes LIKE CONCAT('%[',rd.RegraID,']%') WHERE suser.id = 3531 AND rd.Recursos LIKE '%ContasAReceber%' AND (rd.Unidades LIKE '%|6|%' OR rd.Unidades = '' )
 	end if
@@ -193,10 +216,10 @@ if existePagto="" then
 							ValorDesconto=0
 						end if
 
-
 						if DescontoMaximo < CCUR(ValorDesconto)  then
 							totalValorDescontado = totalValorDescontado + CCUR(ValorDesconto)
 							ValorDesconto = 0
+							PossuiPermissaoParaDesconto=False
 						end if
 					end if
                 end if
@@ -331,14 +354,32 @@ if erro="" then
             db.execute("insert into itensinvoice_bck (`id`, `InvoiceID`, `Tipo`, `Quantidade`, `CategoriaID`, `ItemID`, `ValorUnitario`, `Desconto`, `Descricao`, `Executado`, `DataExecucao`, `HoraExecucao`, `GrupoID`, `AgendamentoID`, `sysUser`, `sysDate`, `ProfissionalID`, `EspecialidadeID`, `HoraFim`, `Acrescimo`, `AtendimentoID`, `Associacao`, `CentroCustoID`, `OdontogramaObj`, `PacoteID`, `DHUp`, `GeradoAutomaticamente`) select `id`, `InvoiceID`, `Tipo`, `Quantidade`, `CategoriaID`, `ItemID`, `ValorUnitario`, `Desconto`, `Descricao`, `Executado`, `DataExecucao`, `HoraExecucao`, `GrupoID`, `AgendamentoID`, `sysUser`, `sysDate`, `ProfissionalID`, `EspecialidadeID`, `HoraFim`, `Acrescimo`, `AtendimentoID`, `Associacao`, `CentroCustoID`, `OdontogramaObj`, `PacoteID`, `DHUp`, `GeradoAutomaticamente` from itensinvoice where InvoiceID="&InvoiceID)
         end if
 
-        sqlExecute = "delete from itensinvoice where InvoiceID="&InvoiceID
+        sqlWhereItens = "ii.InvoiceID="&InvoiceID
+
         if itensStr&""<>"" then
 		    sqlExecute = "delete from itensinvoice where InvoiceID="&InvoiceID&" AND id not in ("&itensStr&")"
-			db.execute("DELETE FROM tissguiasinvoice WHERE InvoiceID="&InvoiceID&" AND ItemInvoiceID not in ("&itensStr&")")			
+			db.execute("DELETE FROM tissguiasinvoice WHERE InvoiceID="&InvoiceID&" AND ItemInvoiceID not in ("&itensStr&")")
+			sqlWhereItens = "ii.InvoiceID="&InvoiceID&" AND ii.id not in ("&itensStr&")"
 		end if
 
-		call gravaLogs(sqlExecute ,"AUTO", "Item excluído manualmente","InvoiceID")
-		db_execute(sqlExecute)
+        set ItensExcluidosSQL = db.execute("SELECT COALESCE(proc.NomeProcedimento, ii.Descricao) Descricao, ii.ValorUnitario, ii.Desconto, if(ii.Executado='S','S','N')Executado "&_
+                                           "FROM itensinvoice ii  "&_
+                                           "LEFT JOIN procedimentos proc on proc.id=ii.ItemID "&_
+                                           "WHERE "&sqlWhereItens)
+
+        if not ItensExcluidosSQL.eof then
+            while not ItensExcluidosSQL.eof
+                DescricaoExclusaoItem = "Descricao: "&ItensExcluidosSQL("Descricao")&" | Valor unit.: R$ "&fn(ItensExcluidosSQL("ValorUnitario"))&" | Desconto: R$ "&fn(ItensExcluidosSQL("Desconto"))&" | Executado: "&ItensExcluidosSQL("Executado")
+                call registraEventoAuditoria("remove_item_fatura", InvoiceID, DescricaoExclusaoItem)
+            ItensExcluidosSQL.movenext
+            wend
+            ItensExcluidosSQL.close
+            set ItensExcluidosSQL=nothing
+            sqlExecute = "delete from itensinvoice where InvoiceID="&InvoiceID
+
+            call gravaLogs(sqlExecute ,"AUTO", "Item excluído manualmente","InvoiceID")
+            db_execute(sqlExecute)
+        end if
 
 		'-> roda de novo o processo de cima
 		totInvo = 0
@@ -410,7 +451,7 @@ if erro="" then
                         valorUnitario="0"
                     end if
 
-					if descontoIgual = False then 
+					if descontoIgual = False then
 						if not rsDescontosUsuario.eof then
 							while not rsDescontosUsuario.eof
 								procedimentoText = rsDescontosUsuario("Procedimentos")
@@ -477,7 +518,7 @@ if erro="" then
 				NewItemID = Row
 			end if
 
-			if temregradesconto=1 then
+			if temregradesconto=1 and not PossuiPermissaoParaDesconto then
 				'Gravar esses dados em outra tabela
 
 				DescontoInput = ref("Desconto"&ii)
@@ -491,6 +532,7 @@ if erro="" then
 
 				if temdescontocadastrado=1 and CCUR(DescontoInput) > 0  then
 					msgExtra = "Alguns itens necessitam de aprovação para o desconto"
+					statusAlertSuccess="warning"
 					set DescontosSQL = db.execute("select * from descontos_pendentes where ItensInvoiceID = "&NewItemID&"")
 					if not DescontosSQL.eof then
 						sqlInsertpendente = "update descontos_pendentes set DataHora=NOW(),Desconto = "&treatvalzero(ref("Desconto"&ii))&", sysUserAutorizado=null,DataHoraAutorizado=null,  Status = 0, SysUser = "&session("User")&" where id = " & DescontosSQL("id")
@@ -600,11 +642,11 @@ if erro="" then
             DescricaoLog=""
         end if
         if scp()=1 then
-			sqlInvoice = "update sys_financialinvoices set Rateado="&contaRatiada&", AccountID="&AccountID&", AssociationAccountID="&AssociationAccountID&", Value="&treatvalzero(ref("Valor"))&", Tax=1, Currency='BRL', Recurrence="&treatvalnull(ref("Recurrence"))&", RecurrenceType='"&ref("RecurrenceType")&"', FormaID="&treatvalzero(splForma(0))&", ContaRectoID="&treatvalzero(splForma(1))&", TabelaID="& treatvalnull(ref("invTabelaID")) &", ProfissionalSolicitante='"&ref("ProfissionalSolicitante")&"', nroNFe='"& ref("nroNFe") &"', CompanyUnitID="&treatvalzero(ref("CompanyUnitID"))&", sysActive=1 "& sqlCaixaID & sqlUsuario & gravaData &" where id="&InvoiceID
+			sqlInvoice = "update sys_financialinvoices set Voucher='"&Voucher&"', Rateado="&contaRateada&", AccountID="&AccountID&", AssociationAccountID="&AssociationAccountID&", Value="&treatvalzero(ref("Valor"))&", Tax=1, Currency='BRL', Recurrence="&treatvalnull(ref("Recurrence"))&", RecurrenceType='"&ref("RecurrenceType")&"', FormaID="&treatvalzero(splForma(0))&", ContaRectoID="&treatvalzero(splForma(1))&", TabelaID="& treatvalnull(ref("invTabelaID")) &", ProfissionalSolicitante='"&ref("ProfissionalSolicitante")&"', nroNFe='"& ref("nroNFe") &"', CompanyUnitID="&treatvalzero(ref("CompanyUnitID"))&", sysActive=1 "& sqlCaixaID & sqlUsuario & gravaData &" where id="&InvoiceID
 			'call gravaLog(sqlInvoice, "AUTO")
 	    	db_execute(sqlInvoice)
         else
-			sqlInvoice = "update sys_financialinvoices set Rateado="&contaRatiada&", AccountID="&AccountID&", AssociationAccountID="&AssociationAccountID&", Value="&treatvalzero(ref("Valor"))&", Tax=1, Currency='BRL', Recurrence="&treatvalnull(ref("Recurrence"))&", RecurrenceType='"&ref("RecurrenceType")&"', FormaID="&treatvalzero(splForma(0))&", ContaRectoID="&treatvalzero(splForma(1))&", TabelaID="& treatvalnull(ref("invTabelaID")) &", ProfissionalSolicitante='"&ref("ProfissionalSolicitante")&"', nroNFe='"& ref("nroNFe") &"', CompanyUnitID="&treatvalzero(ref("CompanyUnitID"))&", sysActive=1 "& sqlCaixaID & sqlUsuario & gravaData &" where id="&InvoiceID
+			sqlInvoice = "update sys_financialinvoices set Voucher='"&Voucher&"', Rateado="&contaRateada&", AccountID="&AccountID&", AssociationAccountID="&AssociationAccountID&", Value="&treatvalzero(ref("Valor"))&", Tax=1, Currency='BRL', Recurrence="&treatvalnull(ref("Recurrence"))&", RecurrenceType='"&ref("RecurrenceType")&"', FormaID="&treatvalzero(splForma(0))&", ContaRectoID="&treatvalzero(splForma(1))&", TabelaID="& treatvalnull(ref("invTabelaID")) &", ProfissionalSolicitante='"&ref("ProfissionalSolicitante")&"', nroNFe='"& ref("nroNFe") &"', CompanyUnitID="&treatvalzero(ref("CompanyUnitID"))&", sysActive=1 "& sqlCaixaID & sqlUsuario & gravaData &" where id="&InvoiceID
 		' 	call gravaLog(sqlInvoice, "AUTO")
 
 			db_execute(sqlInvoice)
@@ -714,9 +756,9 @@ if erro="" then
 	});
 	
 	new PNotify({
-		title: 'Salvo com sucesso!<%=msgExtra%>',
-		text: '',
-		type: 'success',
+		title: 'Salvo com sucesso!',
+		text: '<%=msgExtra%>',
+		type: '<%=statusAlertSuccess%>',
         delay: 1000
 	});
 	geraParcelas('N');
@@ -727,6 +769,7 @@ if erro="" then
         $('#btnSave').prop('disabled', true);
         $('#btnSave').html('AGUARDE...');
         ajxContent('Conta', $('#PacienteID').val(), '1', 'divHistorico');
+        $("#rbtns").fadeOut();
     }else{
         $('#btnSave').prop('disabled', false);
     }
